apps/Pi5/pyproject.toml
[project]
name = "brewbot-pi5"
version = "0.1.0"
description = "Firmware service for BrewBot running on Raspberry Pi 5"
authors = [{ name = "Bartender Boys" }]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.3",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
pi = ["RPi.GPIO>=0.7.1"]
dev = ["pytest>=8.3.0", "ruff>=0.5.0"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

